+++ core/components/babel/lexicon/de/properties.inc.php
@@ -6,7 +6,7 @@
  * @subpackage lexicon
  */
 $_lang['babel.babellinks.activeCls'] = 'CSS-Klasse für den Sprachen-Link der aktiven Sprache.';
-$_lang['babel.babellinks.ignoreSiteStatus'] = 'Gibt an, ob der site_status ignoriert werden soll (wenn die Site offline ist)';
+$_lang['babel.babellinks.ignoreSiteStatus'] = 'Gibt an, ob der site_status ignoriert werden soll (wenn die Site offline ist).';
 $_lang['babel.babellinks.includeUnlinked'] = 'Gibt an, ob ein Kontext, dessen Verknüpfung gelöst wurde, angezeigt werden soll.';
 $_lang['babel.babellinks.resourceId'] = 'ID der Ressource, von der die Links zu Übersetzungen dargestellt werden sollen.';
 $_lang['babel.babellinks.restrictToGroup'] = 'Beschränke die Kontexte in der Snippet-Ausgabe auf die Gruppe des aktuellen Kontexts.';
@@ -17,6 +17,7 @@ $_lang['babel.babellinks.sortdir'] = 'Die Richtung, in die sortiert werden soll.
 $_lang['babel.babellinks.toArray'] = 'Gibt an, ob die Ausgabe stattdessen als Array ausgegeben werden soll.';
 $_lang['babel.babellinks.toPlaceholder'] = 'Gibt an, ob das Ergebnis im Platzhalter mit dem angegebenen Namen ausgegeben werden soll.';
 $_lang['babel.babellinks.tpl'] = 'Chunk für die Darstellung eines Sprachen-Links.';
+$_lang['babel.babellinks.useRequestProperties'] = 'Gibt an, ob die aktuellen Request Eigenschaften in den Links zu den übersetzten Ressourcen verwendet werden.';
 $_lang['babel.babellinks.wrapperTpl'] = 'Chunk, um alle Sprachlinks zu verpacken. Der Platzhalter für alle Links ist [[+babelLinks]].';
 $_lang['babel.babeltranslation.contextKey'] = 'Schlüssel des Kontexts, in der die übersetzte Ressource ermittelt werden soll.';
 $_lang['babel.babeltranslation.cultureKey'] = 'Sprach-Schlüssel, mit dem die übersetzte Ressource ermittelt werden soll. Wird nur verwendet, wenn Kontext Schlüssel nicht angegeben wurde. Wenn beide Schlüssel weggelassen werden, wird der systemweit eingestellte cultureKey verwendet.';