+++ core/components/babel/lexicon/es/languages.inc.php
@@ -1,0 +1,40 @@
+<?php
+/**
+ * Languages Lexicon Entries for Babel
+ *
+ * @package babel
+ * @subpackage lexicon
+ */
+$_lang['babel.language_ar'] = 'Árabe';
+$_lang['babel.language_bg'] = 'Búlgaro';
+$_lang['babel.language_ca'] = 'Catalán';
+$_lang['babel.language_cs'] = 'Checa';
+$_lang['babel.language_da'] = 'Danés';
+$_lang['babel.language_de'] = 'Alemán';
+$_lang['babel.language_en'] = 'Inglés';
+$_lang['babel.language_es'] = 'Español';
+$_lang['babel.language_fa'] = 'Persa';
+$_lang['babel.language_fi'] = 'Finlandés';
+$_lang['babel.language_fr'] = 'Francés';
+$_lang['babel.language_he'] = 'Hebreo';
+$_lang['babel.language_hu'] = 'Húngaro';
+$_lang['babel.language_id'] = 'Indonesio';
+$_lang['babel.language_it'] = 'Italiano';
+$_lang['babel.language_ja'] = 'Japonés';
+$_lang['babel.language_ko'] = 'Coreano';
+$_lang['babel.language_lt'] = 'Lituano';
+$_lang['babel.language_ms'] = 'Malayo';
+$_lang['babel.language_nl'] = 'Holandés';
+$_lang['babel.language_no'] = 'Noruego (Bokmål)';
+$_lang['babel.language_pl'] = 'Polaco';
+$_lang['babel.language_pt'] = 'Portugués';
+$_lang['babel.language_ro'] = 'Rumano';
+$_lang['babel.language_ru'] = 'Ruso';
+$_lang['babel.language_sk'] = 'Eslovaco';
+$_lang['babel.language_sl'] = 'Esloveno';
+$_lang['babel.language_sr'] = 'Serbio';
+$_lang['babel.language_sv'] = 'Sueco';
+$_lang['babel.language_tr'] = 'Turco';
+$_lang['babel.language_uk'] = 'Ucraniano';
+$_lang['babel.language_vi'] = 'Vietnamita';
+$_lang['babel.language_zh'] = 'Chino';