+++ core/components/babel/lexicon/fr/properties.inc.php
@@ -6,20 +6,21 @@
  * @subpackage lexicon
  */
 $_lang['babel.babellinks.activeCls'] = 'Classe CSS pour les liens de la langue active.';
-$_lang['babel.babellinks.ignoreSiteStatus'] = 'Drapeau indiquant s\'il faut ignorer le site_status (lorsque le site est hors ligne)';
-$_lang['babel.babellinks.includeUnlinked'] = 'Indique s\'il faut afficher le contexte non lié.';
+$_lang['babel.babellinks.ignoreSiteStatus'] = 'Indique si le site_status est ignoré (lorsque le site est hors ligne).';
+$_lang['babel.babellinks.includeUnlinked'] = 'Indique si les contextes non liés sont affichés.';
 $_lang['babel.babellinks.resourceId'] = 'ID de la ressource dont les liens vers les traductions doivent être affichés.';
 $_lang['babel.babellinks.restrictToGroup'] = 'Restreindre les contextes dans la sortie du snippet au groupe du contexte actuel.';
-$_lang['babel.babellinks.showCurrent'] = 'Indicateur permettant d\'afficher ou non un lien vers une traduction de la langue actuelle.';
-$_lang['babel.babellinks.showUnpublished'] = 'Indique s\'il faut afficher les traductions non publiées.';
+$_lang['babel.babellinks.showCurrent'] = 'Indique si un lien vers la langue actuelle est affiché.';
+$_lang['babel.babellinks.showUnpublished'] = 'Indique si les traductions non publiées sont affichées.';
 $_lang['babel.babellinks.sortby'] = 'Champ contextuel par lequel trier les résultats.';
 $_lang['babel.babellinks.sortdir'] = 'La direction du tri.';
-$_lang['babel.babellinks.toArray'] = 'Indique s\'il faut afficher la sortie sous forme de tableau.';
-$_lang['babel.babellinks.toPlaceholder'] = 'Indique s\'il faut afficher le résultat dans le nom du placeholder donné.';
+$_lang['babel.babellinks.toArray'] = 'Indique si la sortie est restituée sous la forme d\'un tableau.';
+$_lang['babel.babellinks.toPlaceholder'] = 'Spécifie si la sortie est vidée dans le nom de l\'espace réservé donné.';
 $_lang['babel.babellinks.tpl'] = 'Chunk pour afficher un lien de langue.';
+$_lang['babel.babellinks.useRequestProperties'] = 'Spécifie si les propriétés de la requête en cours sont utilisées dans les liens vers les ressources traduites.';
 $_lang['babel.babellinks.wrapperTpl'] = 'Morceau permettant de regrouper tous les liens linguistiques. L\'espace réservé pour tous les liens est [[+babelLinks]].';
 $_lang['babel.babeltranslation.contextKey'] = 'Clé du contexte dans lequel la ressource traduite doit être déterminée.';
 $_lang['babel.babeltranslation.cultureKey'] = 'Clé de la culture dans laquelle la ressource traduite doit être déterminée. Utilisée uniquement dans le cas où contextKey n\'a pas été spécifié. Si les deux sont omis : utilise la cultureKey actuellement définie.';
 $_lang['babel.babeltranslation.resourceId'] = 'Liste séparée par des virgules d\'ID de ressources dont il convient de déterminer une liste séparée par des virgules de ressources traduites dans la contextKey/cultureKey.';
-$_lang['babel.babeltranslation.showUnpublished'] = 'Indiquer s\'il faut afficher les traductions non publiées.';
-$_lang['babel.babeltranslation.toPlaceholder'] = 'Indique s\'il faut afficher le résultat dans le nom du placeholder donné.';
+$_lang['babel.babeltranslation.showUnpublished'] = 'Indique si les traductions non publiées sont affichées.';
+$_lang['babel.babeltranslation.toPlaceholder'] = 'Indique s\'il faut enregistrer la sortie dans le nom de l\'espace réservé donné.';