+++ core/components/babel/lexicon/it/properties.inc.php
@@ -6,20 +6,21 @@
  * @subpackage lexicon
  */
 $_lang['babel.babellinks.activeCls'] = 'Classe CSS per il link della lingua attiva.';
-$_lang['babel.babellinks.ignoreSiteStatus'] = 'Contrassegno se ignorare il site_status (quando il sito è offline)';
-$_lang['babel.babellinks.includeUnlinked'] = 'Contrassegnare se mostrare il contesto non collegato.';
+$_lang['babel.babellinks.ignoreSiteStatus'] = 'Specifica se il site_status viene ignorato (quando il sito è offline).';
+$_lang['babel.babellinks.includeUnlinked'] = 'Specifica se vengono mostrati i contesti non collegati.';
 $_lang['babel.babellinks.resourceId'] = 'ID della risorsa i cui links alla traduzione dovrebbero essere mostrati.';
 $_lang['babel.babellinks.restrictToGroup'] = 'Limita i contesti nell\'output dello snippet al gruppo del contesto corrente.';
-$_lang['babel.babellinks.showCurrent'] = 'Flag se mostrare un link alla traduzione per la lingua corrente.';
-$_lang['babel.babellinks.showUnpublished'] = 'Flag se mostrare le traduzioni non pubblicate.';
+$_lang['babel.babellinks.showCurrent'] = 'Specifica se viene visualizzato un collegamento alla lingua corrente.';
+$_lang['babel.babellinks.showUnpublished'] = 'Specifica se vengono mostrate le traduzioni non pubblicate.';
 $_lang['babel.babellinks.sortby'] = 'Campo contestuale per cui ordinare i risultati.';
 $_lang['babel.babellinks.sortdir'] = 'La direzione in cui ordinare i risultati.';
-$_lang['babel.babellinks.toArray'] = 'Flag per decidere se scaricare l\'output come array.';
-$_lang['babel.babellinks.toPlaceholder'] = 'Flag che indica se scaricare l\'output nel nome del segnaposto dato.';
+$_lang['babel.babellinks.toArray'] = 'Specifica se l\'output viene scaricato come array.';
+$_lang['babel.babellinks.toPlaceholder'] = 'Specifica se l\'output viene scaricato nel nome del segnaposto dato.';
 $_lang['babel.babellinks.tpl'] = 'Chunk per visualizzare il link della lingua.';
+$_lang['babel.babellinks.useRequestProperties'] = 'Specifica se le proprietà della richiesta corrente sono usate nei collegamenti alle risorse tradotte.';
 $_lang['babel.babellinks.wrapperTpl'] = 'Chunk per avvolgere tutti i collegamenti alle lingue. Il segnaposto per tutti i collegamenti è [[+babelLinks]].';
 $_lang['babel.babeltranslation.contextKey'] = 'Chiave del contesto in cui la risorsa tradotta dovrebbe essere determinata.';
 $_lang['babel.babeltranslation.cultureKey'] = 'Chiave della cultura in cui deve essere determinata la risorsa tradotta. Utilizzato solo nel caso in cui contextKey non sia stato specificato. Se entrambi sono stati omessi, lo snippet è il cultureKey attualmente impostato.';
 $_lang['babel.babeltranslation.resourceId'] = 'Elenco separato da virgole di ID di risorse di cui si deve determinare un elenco separato da virgole di risorse tradotte nel contextKey/cultureKey.';
-$_lang['babel.babeltranslation.showUnpublished'] = 'Flag se mostrare le traduzioni non pubblicate.';
-$_lang['babel.babeltranslation.toPlaceholder'] = 'Flag che indica se scaricare l\'output nel nome del segnaposto dato.';
+$_lang['babel.babeltranslation.showUnpublished'] = 'Specifica se vengono mostrate le traduzioni non pubblicate.';
+$_lang['babel.babeltranslation.toPlaceholder'] = 'Specifica se scaricare l\'output nel nome del segnaposto dato.';