+++ core/components/babel/lexicon/nl/properties.inc.php
@@ -8,20 +8,21 @@
  * @subpackage lexicon
  */
 $_lang['babel.babellinks.activeCls'] = 'CSS class voor de actieve taal link.';
-$_lang['babel.babellinks.ignoreSiteStatus'] = 'Vlag of de site_status moet worden genegeerd (als de site offline is)';
-$_lang['babel.babellinks.includeUnlinked'] = 'Vlag of niet-gelinkte context moet worden weergegeven.';
+$_lang['babel.babellinks.ignoreSiteStatus'] = 'Geeft aan of de site_status wordt genegeerd (als de site offline is).';
+$_lang['babel.babellinks.includeUnlinked'] = 'Geeft aan of de niet-gelinkte context wordt weergegeven.';
 $_lang['babel.babellinks.resourceId'] = 'ID van resource waarvan link naar vertaling moeten worden weergegeven.';
 $_lang['babel.babellinks.restrictToGroup'] = "Beperk de contexten in de knipseluitvoer tot de groep van de huidige context.";
-$_lang['babel.babellinks.showCurrent'] = 'Wel of niet tonen van link naar vertaling van de huidige taal.';
-$_lang['babel.babellinks.showUnpublished'] = 'Wel of niet tonen van ongepubliceerde vertalingen.';
+$_lang['babel.babellinks.showCurrent'] = 'Geeft aan of een link naar de huidige taal wordt weergegeven.';
+$_lang['babel.babellinks.showUnpublished'] = 'Geeft aan of de niet-gepubliceerde vertalingen worden weergegeven.';
 $_lang['babel.babellinks.sortby'] = 'Contextveld om de resultaten op te sorteren.';
 $_lang['babel.babellinks.sortdir'] = 'De richting waarin gesorteerd moet worden.';
-$_lang['babel.babellinks.toArray'] = 'Vlag of de uitvoer als een matrix moet worden gedumpt.';
-$_lang['babel.babellinks.toPlaceholder'] = 'Vlag of de uitvoer in de naam van de opgegeven plaatshouder moet worden gedumpt.';
+$_lang['babel.babellinks.toArray'] = 'Geeft aan of de uitvoer wordt gedumpt als een matrix.';
+$_lang['babel.babellinks.toPlaceholder'] = 'Geeft aan of de uitvoer wordt gedumpt in de naam van de opgegeven plaatshouder.';
 $_lang['babel.babellinks.tpl'] = 'Chunk voor de weergave van een taal link.';
+$_lang['babel.babellinks.useRequestProperties'] = 'Geeft aan of de huidige request eigenschappen worden gebruikt in de links naar de vertaalde bronnen.';
 $_lang['babel.babellinks.wrapperTpl'] = 'Chunk om alle taallinks in te pakken. De plaatshouder voor alle links is [[+babelLinks]].';
 $_lang['babel.babeltranslation.contextKey'] = 'Context key waarin vertaalde bron dient te worden bepaald.';
 $_lang['babel.babeltranslation.cultureKey'] = 'Sleutel van cultuur waarin vertaalde bron moet worden bepaald. Wordt alleen gebruikt als contextKey niet is opgegeven. Als beide weggelaten het knipsel de huidige ingestelde cultureKey.';
 $_lang['babel.babeltranslation.resourceId'] = 'Door komma\'s gescheiden lijst van bron-ID\'s waarvan een door komma\'s gescheiden lijst van vertaalde bronnen in de contextKey/cultureKey moet worden bepaald.';
-$_lang['babel.babeltranslation.showUnpublished'] = 'Wel of niet tonen van ongepubliceerde vertalingen.';
+$_lang['babel.babeltranslation.showUnpublished'] = 'Geeft aan of de niet-gepubliceerde vertalingen worden weergegeven.';
 $_lang['babel.babeltranslation.toPlaceholder'] = 'Vlag of de uitvoer in de naam van de opgegeven plaatshouder moet worden gedumpt.';