+++ core/components/babel/lexicon/pt/properties.inc.php
@@ -8,20 +8,21 @@
  * @subpackage lexicon
  */
 $_lang['babel.babellinks.activeCls'] = 'Classe de CSS para a hiperligação da língua ativa.';
-$_lang['babel.babellinks.ignoreSiteStatus'] = 'Sinalizar se o site_status deve ser ignorado (quando o site está offline)';
-$_lang['babel.babellinks.includeUnlinked'] = 'Indicar se deseja apresentar contextos sem associação.';
+$_lang['babel.babellinks.ignoreSiteStatus'] = 'Especifica se o site_status é ignorado (quando o site está offline).';
+$_lang['babel.babellinks.includeUnlinked'] = 'Especifica se o contexto não ligado é mostrado.';
 $_lang['babel.babellinks.resourceId'] = 'ID da página (recurso) para a qual as hiperligações das traduções devam ser apresentadas.';
 $_lang['babel.babellinks.restrictToGroup'] = 'Restringir os contextos na saída do snippet ao grupo do contexto atual.';
-$_lang['babel.babellinks.showCurrent'] = 'Indicar se deseja apresentar uma hiperligação para uma tradução do idioma atual.';
-$_lang['babel.babellinks.showUnpublished'] = 'Indicar se deseja apresentar traduções não publicadas.';
+$_lang['babel.babellinks.showCurrent'] = 'Especifica se é apresentada uma ligação para a língua atual.';
+$_lang['babel.babellinks.showUnpublished'] = 'Especifica se as traduções não publicadas são mostradas.';
 $_lang['babel.babellinks.sortby'] = 'Campo de contexto para ordenar os resultados por.';
 $_lang['babel.babellinks.sortdir'] = 'A direção para ordenar.';
-$_lang['babel.babellinks.toArray'] = 'Sinaliza se a saída deve ser descarregada como uma matriz.';
-$_lang['babel.babellinks.toPlaceholder'] = 'Sinaliza se a saída deve ser despejada no nome do espaço reservado fornecido.';
+$_lang['babel.babellinks.toArray'] = 'Especifica se a saída é descarregada como uma matriz.';
+$_lang['babel.babellinks.toPlaceholder'] = 'Especifica se a saída é descarregada para o nome do espaço reservado dado.';
 $_lang['babel.babellinks.tpl'] = 'Chunk para apresentar a hiperligação da língua.';
+$_lang['babel.babellinks.useRequestProperties'] = 'Especifica se as propriedades do pedido atual são utilizadas nas ligações para os recursos traduzidos.';
 $_lang['babel.babellinks.wrapperTpl'] = 'Fragmento para envolver todos os links de idiomas. O marcador de posição para todos os links é [[+babelLinks]].';
 $_lang['babel.babeltranslation.contextKey'] = 'Chave do contexto em que a página (recurso) traduzida deva ser determinado.';
 $_lang['babel.babeltranslation.cultureKey'] = 'Chave da cultura na qual o recurso traduzido deve ser determinado. Utilizada apenas no caso de contextKey não ter sido especificada. Se ambos forem omitidos, o snippet usa a cultureKey atualmente definida.';
 $_lang['babel.babeltranslation.resourceId'] = 'Lista separada por vírgulas de IDs de recursos dos quais deve ser determinada uma lista separada por vírgulas de recursos traduzidos na contextKey/cultureKey.';
-$_lang['babel.babeltranslation.showUnpublished'] = 'Indicar se deseja apresentar traduções não publicadas.';
-$_lang['babel.babeltranslation.toPlaceholder'] = 'Sinaliza se a saída deve ser despejada no nome do espaço reservado fornecido.';
+$_lang['babel.babeltranslation.showUnpublished'] = 'Especifica se as traduções não publicadas são mostradas.';
+$_lang['babel.babeltranslation.toPlaceholder'] = 'Especifica se a saída deve ser despejada no nome do espaço reservado fornecido.';