+++ core/components/babel/lexicon/ro/properties.inc.php
@@ -8,20 +8,21 @@
  * @subpackage lexicon
  */
 $_lang['babel.babellinks.activeCls'] = 'Clasa CSS pentru legătura activă către resursele traduse.';
-$_lang['babel.babellinks.ignoreSiteStatus'] = 'Semnalizează dacă se ignoră statusul site-ului (atunci când site-ul este offline)';
-$_lang['babel.babellinks.includeUnlinked'] = 'Semnalizează dacă se afișează sau nu contextul fără legătură.';
+$_lang['babel.babellinks.ignoreSiteStatus'] = 'Specifică dacă site_status este ignorat (atunci când site-ul este offline).';
+$_lang['babel.babellinks.includeUnlinked'] = 'Specifică dacă sunt afișate contextele nelegate.';
 $_lang['babel.babellinks.resourceId'] = 'ID-ul resursei de afișat.';
 $_lang['babel.babellinks.restrictToGroup'] = 'Restrânge contextele din extrasul de fragment la grupul contextului curent.';
-$_lang['babel.babellinks.showCurrent'] = 'Afișează o legătură către limba curentă.';
-$_lang['babel.babellinks.showUnpublished'] = 'Afișează traducerile nepublicate.';
+$_lang['babel.babellinks.showCurrent'] = 'Specifică dacă este afișată o legătură către limba curentă.';
+$_lang['babel.babellinks.showUnpublished'] = 'Specifică dacă sunt afișate traducerile nepublicate.';
 $_lang['babel.babellinks.sortby'] = 'Câmpul contextual după care se sortează rezultatele.';
 $_lang['babel.babellinks.sortdir'] = 'Direcția de sortare.';
-$_lang['babel.babellinks.toArray'] = 'Semnalizează dacă se dorește să se arunce ieșirea sub formă de matrice.';
-$_lang['babel.babellinks.toPlaceholder'] = 'Semnalizează dacă trebuie să se arunce ieșirea în numele de placeholder dat.';
+$_lang['babel.babellinks.toArray'] = 'Specifică dacă ieșirea este descărcată ca un array.';
+$_lang['babel.babellinks.toPlaceholder'] = 'Specifică dacă rezultatul este aruncat în numele plasatorului dat.';
 $_lang['babel.babellinks.tpl'] = 'Elementul chunk ce afișează legăturile către resursele traduse.';
+$_lang['babel.babellinks.useRequestProperties'] = 'Specifică dacă proprietățile cererii curente sunt utilizate în legăturile către resursele traduse.';
 $_lang['babel.babellinks.wrapperTpl'] = 'Bucată pentru a înfășura toate legăturile de limbă. Locul pentru toate legăturile este [[+babelLinks]].';
 $_lang['babel.babeltranslation.contextKey'] = 'Cheia de context a resursei de tradus.';
 $_lang['babel.babeltranslation.cultureKey'] = 'Cheia de cultură în care ar trebui să fie determinată resursa tradusă. Se utilizează numai în cazul în care nu s-a specificat contextKey. În cazul în care ambele sunt omise, se utilizează cultureKey setată în prezent.';
 $_lang['babel.babeltranslation.resourceId'] = 'Listă de ID-uri de resurse separate prin virgulă din care ar trebui determinată o listă de resurse traduse separate prin virgulă în contextKey/cultureKey.';
-$_lang['babel.babeltranslation.showUnpublished'] = 'Afișează traducerile nepublicate.';
-$_lang['babel.babeltranslation.toPlaceholder'] = 'Semnalizează dacă trebuie să se arunce ieșirea în numele de placeholder dat.';
+$_lang['babel.babeltranslation.showUnpublished'] = 'Specifică dacă sunt afișate traducerile nepublicate.';
+$_lang['babel.babeltranslation.toPlaceholder'] = 'Specifică dacă se va arunca rezultatul în numele titularului de poziție dat.';